Tuning compiler thread count for optimal application install times
dex2oat decides the number of compiler threads to be spawned based
on the number of cores in the system. Using this max possible compiler
threads causes over parallelization on some platforms and there by
leads to longer install times. Reducing to fewer threads on these
platforms, improves this time by 60% to 100%

Change-Id: Iea44d892e20fd3f2c92565300dfa5f96714bc5b4
CRs-Fixed: 818615
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>
Signed-off-by: Louis Popi <theh2o64@gmail.com>